requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-environ==0.4.5
gunicorn==20.0.4
idna==2.10
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
